third_party/WebKit/LayoutTests/fast/xmlhttprequest/instanceof-XMLHttpRequest.html

<!DOCTYPE HTML>

<body>

<script src="../../resources/js-test.js"></script>

<script>
description("instanceof XMLHttpRequest test");

shouldBeTrue("(new XMLHttpRequest()) instanceof XMLHttpRequest");
shouldBeFalse("(new Array()) instanceof XMLHttpRequest");

shouldBeTrue("(new XMLHttpRequest()) instanceof XMLHttpRequestEventTarget");
shouldBeFalse("(new Array()) instanceof XMLHttpRequestEventTarget");

shouldBeTrue("(new XMLHttpRequest).upload instanceof XMLHttpRequestUpload");
shouldBeTrue("(new XMLHttpRequest).upload instanceof XMLHttpRequestEventTarget");
</script>
</body>
