third_party/blink/web_tests/fast/xmlhttprequest/instanceof-XMLHttpRequest-expected.txt

instanceof XMLHttpRequest test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS (new XMLHttpRequest()) instanceof XMLHttpRequest is true
PASS (new Array()) instanceof XMLHttpRequest is false
PASS (new XMLHttpRequest()) instanceof XMLHttpRequestEventTarget is true
PASS (new Array()) instanceof XMLHttpRequestEventTarget is false
PASS (new XMLHttpRequest).upload instanceof XMLHttpRequestUpload is true
PASS (new XMLHttpRequest).upload instanceof XMLHttpRequestEventTarget is true
PASS successfullyParsed is true

TEST COMPLETE
